Java language lexer and parser added
A lexer in `flex` and parser in `bison`, for the
Java language, have been added. It is able to
parse a simple Java program and count the number of class and function definitions, as
well as the number of function calls.
TODO:
- Generate an abstract syntax tree (AST) from
the parser.
- Convert the AST to LLVM-IR intermediate
language, so it can be assembled to native machine language.